feat: carousel 구조 작성
